+++ HISTORY.md
@@ -5,7 +5,7 @@
 - Add `--reset-cache` to delete cached vulnerability data.
 - Add `--ignore-cache` to temporarily ignore cached vulnerability data.
 - Handle invalid JSON caches (by ignoring/overwriting them)
-
+- Allow to ignore vulnerabilities by Sonatype ID or CVE ID 
 
 ## 0.5.0